Add cl2703 as a python module to the ament index
This change exposes a new Python3 module, `cl2703`, through the ament
index and allows the modules under cl2703/scripts to be imported from it.

For example, the rack shifting utilities from Task 2B can be used after
something like `import cl2703.task2b`

This is useful for scripts in other packages; the ebot docking controller
is an example of this.